refactor(messaging): remove deprecated namespaced API
