+++ packages/app/lib/common/index.ts
@@ -422,47 +422,6 @@ const mapOfDeprecationReplacements: DeprecationMap = {
       HttpsErrorCode: 'HttpsErrorCode',
     },
   },
-  messaging: {
-    default: {
-      isAutoInitEnabled: 'isAutoInitEnabled()',
-      isDeviceRegisteredForRemoteMessages: 'isDeviceRegisteredForRemoteMessages()',
-      isNotificationDelegationEnabled: 'isNotificationDelegationEnabled()',
-      isDeliveryMetricsExportToBigQueryEnabled: 'isDeliveryMetricsExportToBigQueryEnabled()',
-      setAutoInitEnabled: 'setAutoInitEnabled()',
-      getInitialNotification: 'getInitialNotification()',
-      getDidOpenSettingsForNotification: 'getDidOpenSettingsForNotification()',
-      getIsHeadless: 'getIsHeadless()',
-      onNotificationOpenedApp: 'onNotificationOpenedApp()',
-      onTokenRefresh: 'onTokenRefresh()',
-      requestPermission: 'requestPermission()',
-      registerDeviceForRemoteMessages: 'registerDeviceForRemoteMessages()',
-      unregisterDeviceForRemoteMessages: 'unregisterDeviceForRemoteMessages()',
-      getAPNSToken: 'getAPNSToken()',
-      setAPNSToken: 'setAPNSToken()',
-      hasPermission: 'hasPermission()',
-      onDeletedMessages: 'onDeletedMessages()',
-      onMessageSent: 'onMessageSent()',
-      onSendError: 'onSendError()',
-      setBackgroundMessageHandler: 'setBackgroundMessageHandler()',
-      setOpenSettingsForNotificationsHandler: 'setOpenSettingsForNotificationsHandler()',
-      sendMessage: 'sendMessage()',
-      subscribeToTopic: 'subscribeToTopic()',
-      unsubscribeFromTopic: 'unsubscribeFromTopic()',
-      setNotificationDelegationEnabled: 'setNotificationDelegationEnabled()',
-      // Actual firebase-js-sdk methods
-      getToken: 'getToken()',
-      deleteToken: 'deleteToken()',
-      onMessage: 'onMessage()',
-      isSupported: 'isSupported()',
-      setDeliveryMetricsExportToBigQuery:
-        'experimentalSetDeliveryMetricsExportedToBigQueryEnabled()',
-    },
-    statics: {
-      AuthorizationStatus: 'AuthorizationStatus',
-      NotificationAndroidPriority: 'NotificationAndroidPriority',
-      NotificationAndroidVisibility: 'NotificationAndroidVisibility',
-    },
-  },
   perf: {
     default: {
       setPerformanceCollectionEnabled: 'initializePerformance()',
