+++ packages/messaging/lib/types/messaging.ts
@@ -16,11 +16,6 @@
  */
 
 import type { ReactNativeFirebase } from '@react-native-firebase/app';
-import {
-  AuthorizationStatus as AuthorizationStatusConst,
-  NotificationAndroidPriority as NotificationAndroidPriorityConst,
-  NotificationAndroidVisibility as NotificationAndroidVisibilityConst,
-} from '../statics';
 
 // ============ Types ============
 
@@ -726,70 +721,3 @@ export interface Messaging extends ReactNativeFirebase.FirebaseModule {
    */
   isSupported(): Promise<boolean>;
 }
-
-// ============ Statics Interface ============
-
-/**
- * Static properties available on firebase.messaging
- */
-
-export interface Statics {
-  SDK_VERSION: string;
-  /** @deprecated See {@link https://github.com/invertase/react-native-firebase/issues/6283 #6283}. */
-  AuthorizationStatus: typeof AuthorizationStatusConst;
-  NotificationAndroidPriority: typeof NotificationAndroidPriorityConst;
-  NotificationAndroidVisibility: typeof NotificationAndroidVisibilityConst;
-}
-
-// ============ Module Augmentation ============
-/* eslint-disable @typescript-eslint/no-namespace */
-declare module '@react-native-firebase/app' {
-  namespace ReactNativeFirebase {
-    interface Module {
-      messaging: FirebaseModuleWithStaticsAndApp<Messaging, Statics>;
-    }
-    interface FirebaseApp {
-      messaging(): Messaging;
-    }
-  }
-}
-
-// ============ Backwards Compatibility Namespace ============
-
-/* eslint-disable @typescript-eslint/no-namespace */
-type _Messaging = Messaging;
-type _MessagingStatics = Statics;
-type _RemoteMessage = RemoteMessage;
-type _MessagePriority = MessagePriority;
-type _FcmOptions = FcmOptions;
-type _NativeTokenOptions = NativeTokenOptions;
-type _GetTokenOptions = GetTokenOptions;
-type _Notification = Notification;
-type _NotificationPayload = NotificationPayload;
-type _NotificationIOSCriticalSound = NotificationIOSCriticalSound;
-type _NotificationAndroidPriority = NotificationAndroidPriority;
-type _NotificationAndroidVisibility = NotificationAndroidVisibility;
-type _IOSPermissions = IOSPermissions;
-type _AuthorizationStatus = AuthorizationStatus;
-type _SendErrorEvent = SendErrorEvent;
-
-export namespace FirebaseMessagingTypes {
-  export type Module = _Messaging;
-  export type Statics = _MessagingStatics;
-  export type RemoteMessage = _RemoteMessage;
-  export type MessagePriority = _MessagePriority;
-  export type FcmOptions = _FcmOptions;
-  export type NativeTokenOptions = _NativeTokenOptions;
-  export type GetTokenOptions = _GetTokenOptions;
-  export type Notification = _Notification;
-  export type NotificationPayload = _NotificationPayload;
-  export type NotificationIOSCriticalSound = _NotificationIOSCriticalSound;
-  export type NotificationAndroidPriority = _NotificationAndroidPriority;
-  export type NotificationAndroidVisibility = _NotificationAndroidVisibility;
-  /** @deprecated See {@link https://github.com/invertase/react-native-firebase/issues/6283 #6283}. */
-  export type IOSPermissions = _IOSPermissions;
-  /** @deprecated See {@link https://github.com/invertase/react-native-firebase/issues/6283 #6283}. */
-  export type AuthorizationStatus = _AuthorizationStatus;
-  export type SendErrorEvent = _SendErrorEvent;
-}
-/* eslint-enable @typescript-eslint/no-namespace */
