+++ packages/messaging/type-test.ts
@@ -1,6 +1,5 @@
-import messaging, {
-  firebase,
-  FirebaseMessagingTypes,
+import { getApp } from '@react-native-firebase/app';
+import {
   getMessaging,
   deleteToken,
   getToken,
@@ -34,212 +33,21 @@ import messaging, {
   AuthorizationStatus,
   NotificationAndroidPriority,
   NotificationAndroidVisibility,
+  SDK_VERSION,
   type Messaging,
+  type RemoteMessage,
+  type SendErrorEvent,
 } from '.';
+import type { AuthorizationStatus as MessagingAuthStatus } from './lib/types/messaging';
 
-console.log(messaging().app);
-
-// checks module exists at root
-console.log(firebase.messaging().app.name);
-console.log(firebase.messaging().isAutoInitEnabled);
-console.log(firebase.messaging().isDeviceRegisteredForRemoteMessages);
-console.log(firebase.messaging().isNotificationDelegationEnabled);
-console.log(firebase.messaging().isDeliveryMetricsExportToBigQueryEnabled);
-
-// checks module exists at app level
-console.log(firebase.app().messaging().app.name);
-console.log(firebase.app().messaging().isAutoInitEnabled);
-
-const messagingInstance2: Messaging = firebase.messaging();
-console.log(messagingInstance2.app.name);
-
-// checks statics exist
-console.log(firebase.messaging.SDK_VERSION);
-console.log(firebase.messaging.AuthorizationStatus.AUTHORIZED);
-console.log(firebase.messaging.AuthorizationStatus.DENIED);
-console.log(firebase.messaging.AuthorizationStatus.NOT_DETERMINED);
-console.log(firebase.messaging.AuthorizationStatus.PROVISIONAL);
-console.log(firebase.messaging.NotificationAndroidPriority.PRIORITY_LOW);
-console.log(firebase.messaging.NotificationAndroidPriority.PRIORITY_HIGH);
-console.log(firebase.messaging.NotificationAndroidVisibility.VISIBILITY_PRIVATE);
-console.log(firebase.messaging.NotificationAndroidVisibility.VISIBILITY_PUBLIC);
-
-// checks statics exist on defaultExport
-console.log(messaging.firebase.SDK_VERSION);
-
-// checks root exists
-console.log(firebase.SDK_VERSION);
-
-// checks default export supports app arg
-console.log(messaging().app.name);
-
-// checks Module instance APIs
-const messagingInstance = firebase.messaging();
-console.log(messagingInstance.app.name);
-console.log(messagingInstance.isAutoInitEnabled);
-console.log(messagingInstance.isDeviceRegisteredForRemoteMessages);
-console.log(messagingInstance.isNotificationDelegationEnabled);
-console.log(messagingInstance.isDeliveryMetricsExportToBigQueryEnabled);
-
-messagingInstance.setAutoInitEnabled(false).then(() => {
-  console.log('Auto init disabled');
-});
-
-messagingInstance
-  .getInitialNotification()
-  .then((message: FirebaseMessagingTypes.RemoteMessage | null) => {
-    if (message) {
-      console.log(message.data);
-    }
-  });
-
-messagingInstance.getDidOpenSettingsForNotification().then((opened: boolean) => {
-  console.log(opened);
-});
-
-messagingInstance.getIsHeadless().then((isHeadless: boolean) => {
-  console.log(isHeadless);
-});
-
-messagingInstance.getToken().then((token: string) => {
-  console.log(token);
-});
-
-messagingInstance.getToken({ appName: 'test', senderId: 'test-sender' }).then((token: string) => {
-  console.log(token);
-});
-
-messagingInstance.deleteToken().then(() => {
-  console.log('Token deleted');
-});
-
-messagingInstance.deleteToken({ appName: 'test', senderId: 'test-sender' }).then(() => {
-  console.log('Token deleted with options');
-});
-
-const unsubscribeOnMessage = messagingInstance.onMessage(
-  (message: FirebaseMessagingTypes.RemoteMessage) => {
-    console.log(message.data);
-  },
-);
-unsubscribeOnMessage();
-
-const unsubscribeOnNotificationOpenedApp = messagingInstance.onNotificationOpenedApp(
-  (message: FirebaseMessagingTypes.RemoteMessage) => {
-    console.log(message.data);
-  },
-);
-unsubscribeOnNotificationOpenedApp();
-
-const unsubscribeOnTokenRefresh = messagingInstance.onTokenRefresh((token: string) => {
-  console.log(token);
-});
-unsubscribeOnTokenRefresh();
-
-messagingInstance.requestPermission().then((status: FirebaseMessagingTypes.AuthorizationStatus) => {
-  console.log(status);
-});
-
-messagingInstance
-  .requestPermission({
-    alert: true,
-    badge: true,
-    sound: true,
-  })
-  .then((status: FirebaseMessagingTypes.AuthorizationStatus) => {
-    console.log(status);
-  });
-
-messagingInstance.registerDeviceForRemoteMessages().then(() => {
-  console.log('Device registered');
-});
-
-messagingInstance.unregisterDeviceForRemoteMessages().then(() => {
-  console.log('Device unregistered');
-});
-
-messagingInstance.getAPNSToken().then((token: string | null) => {
-  console.log(token);
-});
-
-messagingInstance.setAPNSToken('test-token', 'prod').then(() => {
-  console.log('APNS token set');
-});
-
-messagingInstance.hasPermission().then((status: FirebaseMessagingTypes.AuthorizationStatus) => {
-  console.log(status);
-});
-
-const unsubscribeOnDeletedMessages = messagingInstance.onDeletedMessages(() => {
-  console.log('Messages deleted');
-});
-unsubscribeOnDeletedMessages();
-
-const unsubscribeOnMessageSent = messagingInstance.onMessageSent((messageId: string) => {
-  console.log(messageId);
-});
-unsubscribeOnMessageSent();
-
-const unsubscribeOnSendError = messagingInstance.onSendError(
-  (evt: FirebaseMessagingTypes.SendErrorEvent) => {
-    console.log(evt.messageId);
-    console.log(evt.error);
-  },
-);
-unsubscribeOnSendError();
-
-messagingInstance.setBackgroundMessageHandler(
-  async (message: FirebaseMessagingTypes.RemoteMessage) => {
-    console.log(message.data);
-    return Promise.resolve();
-  },
-);
-
-messagingInstance.setOpenSettingsForNotificationsHandler(
-  (message: FirebaseMessagingTypes.RemoteMessage) => {
-    console.log(message.data);
-  },
-);
-
-messagingInstance
-  .sendMessage({
-    data: { key: 'value' },
-    notification: {
-      title: 'Test',
-      body: 'Test body',
-    },
-    fcmOptions: {},
-  })
-  .then(() => {
-    console.log('Message sent');
-  });
-
-messagingInstance.subscribeToTopic('test-topic').then(() => {
-  console.log('Subscribed to topic');
-});
-
-messagingInstance.unsubscribeFromTopic('test-topic').then(() => {
-  console.log('Unsubscribed from topic');
-});
-
-messagingInstance.setDeliveryMetricsExportToBigQuery(true).then(() => {
-  console.log('Delivery metrics enabled');
-});
-
-messagingInstance.setNotificationDelegationEnabled(true).then(() => {
-  console.log('Notification delegation enabled');
-});
-
-messagingInstance.isSupported().then((supported: boolean) => {
-  console.log(supported);
-});
-
-// checks modular API functions
 const modularMessaging1 = getMessaging();
 console.log(modularMessaging1.app.name);
 
-const modularMessaging2 = getMessaging(firebase.app());
+const modularMessaging2 = getMessaging(getApp());
 console.log(modularMessaging2.app.name);
+
+const typedMessaging: Messaging = modularMessaging1;
+console.log(typedMessaging.app.name);
 
 console.log(isAutoInitEnabled(modularMessaging1));
 
@@ -247,13 +55,11 @@ setAutoInitEnabled(modularMessaging1, false).then(() => {
   console.log('Modular auto init disabled');
 });
 
-getInitialNotification(modularMessaging1).then(
-  (message: FirebaseMessagingTypes.RemoteMessage | null) => {
-    if (message) {
-      console.log(message.data);
-    }
-  },
-);
+getInitialNotification(modularMessaging1).then((message: RemoteMessage | null) => {
+  if (message) {
+    console.log(message.data);
+  }
+});
 
 getDidOpenSettingsForNotification(modularMessaging1).then((opened: boolean) => {
   console.log(opened);
@@ -279,17 +85,14 @@ deleteToken(modularMessaging1, { appName: 'test', senderId: 'test-sender' }).the
   console.log('Modular token deleted with options');
 });
 
-const modularUnsubscribeOnMessage = onMessage(
-  modularMessaging1,
-  (message: FirebaseMessagingTypes.RemoteMessage) => {
-    console.log(message.data);
-  },
-);
+const modularUnsubscribeOnMessage = onMessage(modularMessaging1, (message: RemoteMessage) => {
+  console.log(message.data);
+});
 modularUnsubscribeOnMessage();
 
 const modularUnsubscribeOnNotificationOpenedApp = onNotificationOpenedApp(
   modularMessaging1,
-  (message: FirebaseMessagingTypes.RemoteMessage) => {
+  (message: RemoteMessage) => {
     console.log(message.data);
   },
 );
@@ -300,7 +103,7 @@ const modularUnsubscribeOnTokenRefresh = onTokenRefresh(modularMessaging1, (toke
 });
 modularUnsubscribeOnTokenRefresh();
 
-requestPermission(modularMessaging1).then((status: FirebaseMessagingTypes.AuthorizationStatus) => {
+requestPermission(modularMessaging1).then((status: MessagingAuthStatus) => {
   console.log(status);
 });
 
@@ -308,7 +111,7 @@ requestPermission(modularMessaging1, {
   alert: true,
   badge: true,
   sound: true,
-}).then((status: FirebaseMessagingTypes.AuthorizationStatus) => {
+}).then((status: MessagingAuthStatus) => {
   console.log(status);
 });
 
@@ -330,7 +133,7 @@ setAPNSToken(modularMessaging1, 'modular-test-token', 'sandbox').then(() => {
   console.log('Modular APNS token set');
 });
 
-hasPermission(modularMessaging1).then((status: FirebaseMessagingTypes.AuthorizationStatus) => {
+hasPermission(modularMessaging1).then((status: MessagingAuthStatus) => {
   console.log(status);
 });
 
@@ -346,7 +149,7 @@ modularUnsubscribeOnMessageSent();
 
 const modularUnsubscribeOnSendError = onSendError(
   modularMessaging1,
-  (evt: FirebaseMessagingTypes.SendErrorEvent) => {
+  (evt: SendErrorEvent) => {
     console.log(evt.messageId);
     console.log(evt.error);
   },
@@ -355,18 +158,15 @@ modularUnsubscribeOnSendError();
 
 setBackgroundMessageHandler(
   modularMessaging1,
-  async (message: FirebaseMessagingTypes.RemoteMessage) => {
+  async (message: RemoteMessage) => {
     console.log(message.data);
     return Promise.resolve();
   },
 );
 
-setOpenSettingsForNotificationsHandler(
-  modularMessaging1,
-  (message: FirebaseMessagingTypes.RemoteMessage) => {
-    console.log(message.data);
-  },
-);
+setOpenSettingsForNotificationsHandler(modularMessaging1, (message: RemoteMessage) => {
+  console.log(message.data);
+});
 
 sendMessage(modularMessaging1, {
   data: { modularKey: 'modularValue' },
@@ -403,7 +203,9 @@ experimentalSetDeliveryMetricsExportedToBigQueryEnabled(modularMessaging1, true)
   console.log('Modular delivery metrics enabled');
 });
 
-// checks modular statics exports
 console.log(AuthorizationStatus.AUTHORIZED);
 console.log(NotificationAndroidPriority.PRIORITY_DEFAULT);
 console.log(NotificationAndroidVisibility.VISIBILITY_PUBLIC);
+
+const sdkVersion: string = SDK_VERSION;
+console.log(sdkVersion);
